Fixed Shopping List Addition and removal except missing first ingredient
Main things done were changing the DB calls to a transaction and
figuring out when to drop which table.

One other bug besides first ingredient missing is that first time
recipe is added to shopping list it does not work.

Summary of errors: First time inputting data always fails. Probably
something to do with setting up the first view for the list.